Fix dist/echo.js not being transpiled to ES5
Rollup is skipping plugin step babel()
so the resulting distribution is in
ES6 rather than ES5.

i.e., older browsers like IE11 are
unable to parse this JS.

This required upgrading to the newer
Babel 7 plugin ecosystem so the
'extension' option can be used to
explicitly transpile the echo.ts files.